requirements.txt
# Веб-фреймворк
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# Парсинг документов
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
chardet==5.2.0

# NLP для русского языка
transformers>=4.35.2
torch>=2.0.0
natasha>=1.6.0
razdel>=0.5.0
pymorphy3>=1.3.0
yargy>=0.16.0

# Обработка данных
pandas>=2.1.0
openpyxl>=3.1.0
numpy>=1.26.0

# Утилиты
tqdm==4.66.1
python-dateutil==2.8.2

